Fixed line endings

+++ IState.java
@@ -2,7 +2,7 @@ import java.awt.event.KeyEvent;
 
 interface IState {
     void Init(GraphicsClass graphics);
-    void Update(float elapsedTime);
+    void Update(double elapsedTime);
     void Render(GraphicsClass graphics);
     void OnEnter(String[] params);
     void OnExit();
